# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.BaseDTO:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  dto.CategoryDTO:
    properties:
      createTime:
        type: integer
      description:
        type: string
      fullPath:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      password:
        type: string
      priority:
        type: integer
      slug:
        type: string
      thumbnail:
        type: string
      type:
        type: integer
    type: object
...
  param.Category:
    properties:
      description:
        maxLength: 100
        minLength: 0
        type: string
      id:
        minimum: 0
        type: integer
      name:
        maxLength: 255
        minLength: 1
        type: string
      parentId:
        minimum: 0
        type: integer
      password:
        maxLength: 255
        minLength: 0
        type: string
      priority:
        minimum: 0
        type: integer
      slug:
        maxLength: 255
        minLength: 0
        type: string
      thumbnail:
        maxLength: 1023
        minLength: 0
        type: string
    type: object
...
  vo.CategoryVO:
    properties:
      children:
        items:
          $ref: '#/definitions/vo.CategoryVO'
        type: array
      createTime:
        type: integer
      description:
        type: string
      fullPath:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      password:
        type: string
      priority:
        type: integer
      slug:
        type: string
      thumbnail:
        type: string
      type:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    AirPress 博客系统后端 API 文档,通过 swag 注释驱动自动生成。
    包含后台管理(/api/admin)与前台内容(/api/content)两组接口。
  title: AirPress API
  version: "1.0"
paths:
...
  /admin/categories:
    get:
      consumes:
      - application/json
      description: 支持按 more 参数返回带文章数的详情列表或精简列表,支持排序
      parameters:
      - collectionFormat: multi
        description: 排序字段,如 priority,asc
        in: query
        items:
          type: string
        name: sort
        type: array
      - description: true返回带文章数的详情
        in: query
        name: more
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CategoryDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 查询所有分类
      tags:
      - Admin.Category
    post:
      consumes:
      - application/json
      description: 创建一个新分类,返回创建后的详情
      parameters:
      - description: 分类参数
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/param.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 创建分类
      tags:
      - Admin.Category
  /admin/categories/{categoryID}:
    delete:
      description: 根据分类ID删除指定分类
      parameters:
      - description: 分类ID
        example: 1
        in: path
        name: categoryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 删除分类
      tags:
      - Admin.Category
    get:
      description: 返回指定 ID 的分类详情
      parameters:
      - description: 分类ID
        example: 1
        in: path
        name: categoryID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 根据ID获取分类
      tags:
      - Admin.Category
    put:
      consumes:
      - application/json
      description: 根据分类ID更新分类信息
      parameters:
      - description: 分类ID
        example: 1
        in: path
        name: categoryID
        required: true
        type: integer
      - description: 分类参数
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/param.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 更新分类
      tags:
      - Admin.Category
  /admin/categories/batch:
    put:
      consumes:
      - application/json
      description: 根据分类参数列表批量更新分类
      parameters:
      - description: 分类参数列表
        in: body
        name: categories
        required: true
        schema:
          items:
            $ref: '#/definitions/param.Category'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CategoryDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 批量更新分类
      tags:
      - Admin.Category
  /admin/categories/tree_view:
    get:
      consumes:
      - application/json
      description: 返回所有分类的树形结构
      parameters:
      - collectionFormat: multi
        description: 排序字段,如 priority,asc
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/vo.CategoryVO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 树形分类列表
      tags:
      - Admin.Category
...
securityDefinitions:
  AdminApiKey:
    in: header
    name: Admin-Authorization
    type: apiKey
  ContentApiKey:
    in: header
    name: Content-Authorization
    type: apiKey
swagger: "2.0"